pyproject.toml
[tool.poetry]
name = "modbus-meter-emulator"
version = "0.1.0"
description = ""

authors = ["Sander Bos <mail@sanderbos.net>"]
readme = "README.md"
packages = [{ include = "modbus_meter_emulator" }]

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^2.1.0"
pyyaml = "^6.0.2"
pydantic = "^2.10.4"
pymodbus = {extras = ["serial"], version = "^3.8.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
